+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -95,7 +95,7 @@ import org.apache.logging.log4j.Logger
     clientSideOnly = true,
     useMetadata = true,
     guiFactory = "at.hannibal2.skyhanni.config.ConfigGuiForgeInterop",
-    version = "0.18.Beta.15",
+    version = "0.18.Beta.17",
 )
 class SkyHanniMod {
     @Mod.EventHandler
@@ -296,6 +296,7 @@ class SkyHanniMod {
         loadModule(SlayerItemsOnGround())
         loadModule(DetectBrokenHyperion())
         loadModule(RestorePieceOfWizardPortalLore())
+        loadModule(QuickModMenuSwitch)
 
         init()