+++ src/main/java/at/hannibal2/skyhanni/config/Storage.java
@@ -154,6 +154,7 @@ public class Storage {
 
             @Expose
             public PlotIcon plotIcon = new PlotIcon();
+
             public static class PlotIcon {
                 @Expose
                 public Map<Integer, String> plotList = new HashMap<>();
@@ -191,6 +192,45 @@ public class Storage {
                 @Expose
                 public Map<FarmingItems, ItemStack> farmingItems = new HashMap<>();
             }
+
+            @Expose
+            public long composterEmptyTime = 0;
+
+            @Expose
+            public long lastComposterEmptyWarningTime = 0;
         }
+
+        @Expose
+        public long nextCityProjectParticipationTime = 0L;
+
+        @Expose
+        public Map<String, SlayerProfitList> slayerProfitData = new HashMap<>();
+
+        public static class SlayerProfitList {
+
+            @Expose
+            public Map<String, SlayerItemProfit> items = new HashMap<>();
+
+            @Expose
+            public long mobKillCoins = 0;
+
+            @Expose
+            public long slayerSpawnCost = 0;
+
+            @Expose
+            public int slayerCompletedCount = 0;
+
+            public static class SlayerItemProfit {
+                @Expose
+                public String internalName;
+                @Expose
+                public long timesDropped;
+                @Expose
+                public long totalAmount;
+                @Expose
+                public boolean hidden;
+            }
+        }
+
     }
 }